Abstract common logic for resource hints preload/prefetch into relations/HtmlResourceHint
