+++ lib/relations/HtmlPreloadLink.js
@@ -4,7 +4,7 @@
 
 var util = require('util'),
     extendWithGettersAndSetters = require('../util/extendWithGettersAndSetters'),
-    HtmlRelation = require('./HtmlRelation');
+    HtmlResourceHint = require('./HtmlResourceHint');
 
 function getExtension(url) {
     var matches = url.match(/\.([^$#\?]+)/);
@@ -34,104 +34,12 @@ function getPreloadLinkNode(relation, htmlAsset) {
 }
 
 function HtmlPreloadLink(config) {
-    if (!config.to || !config.to.url) {
-        throw new Error('HtmlPreloadLink: The `to` asset must have a url');
-    }
-
-    HtmlRelation.call(this, config);
+    HtmlResourceHint.call(this, config);
 }
 
-util.inherits(HtmlPreloadLink, HtmlRelation);
+util.inherits(HtmlPreloadLink, HtmlResourceHint);
 
 extendWithGettersAndSetters(HtmlPreloadLink.prototype, {
-    get href() {
-        return this.node.getAttribute('href');
-    },
-
-    set href(href) {
-        this.node.setAttribute('href', href);
-    },
-
-    // Map Assetgraph relations to request destination (<link as="${destination}">).
-    // See https://fetch.spec.whatwg.org/#concept-request-destination
-    get requestDestination() {
-        if (this.to.isImage) {
-            return 'image';
-        }
-
-        if (this.to.type === 'JavaScript') {
-            return 'script';
-        }
-
-        if (this.to.type === 'Css') {
-            return 'style';
-        }
-
-        if (this.to.incomingRelations) {
-            var relType = this.to.incomingRelations[0].type;
-
-            if (['HtmlStyle', 'CssImport'].indexOf(relType) !== -1) {
-                return 'style';
-            }
-
-            if (this.to.type === 'Html' && (relType === 'HtmlFrame' || relType === 'HtmlIFrame')) {
-                return 'document';
-            }
-
-            if (relType === 'HtmlAudio') {
-                return 'media';
-            }
-
-            if (relType === 'HtmlVideo') {
-                return 'media';
-            }
-
-            if (relType === 'CssFontFaceSrc') {
-                return 'font';
-            }
-
-            if (relType === 'HtmlEmbed') {
-                return 'embed';
-            }
-
-            if (relType === 'HtmlObject') {
-                return 'object';
-            }
-        }
-
-        // If the asset hasn't been populated, fall back to best guess based on file extension
-        if (this.to.url) {
-            var extension = getExtension(this.to.url);
-
-            switch (extension) {
-            case 'css':
-                return 'style';
-            case 'js':
-                return 'script';
-            case 'svg':
-            case 'jpg':
-            case 'jpeg':
-            case 'png':
-            case 'gif':
-            case 'webp':
-            case 'ico':
-            case 'tiff':
-            case 'bmp':
-                return 'image';
-            case 'html':
-                return 'document';
-            case 'woff':
-            case 'woff2':
-            case 'ttf':
-            case 'eot':
-            case 'otf':
-                return 'font';
-            }
-        }
-
-        return '';
-    },
-
     get contentType() {
         if ('_contentType' in this) {
             return this._contentType;
@@ -156,44 +64,11 @@ extendWithGettersAndSetters(HtmlPreloadLink.prototype, {
         }
     },
 
-    get as() {
-        if (typeof this._as === 'undefined') {
-            this._as = this.requestDestination;
-        }
-
-        return this._as;
-    },
-
-    set as(type) {
-        var validRequestDestinations = [
-            '',
-            'media',
-            'script',
-            'style',
-            'font',
-            'image',
-            'worker',
-            'embed',
-            'object',
-            'document'
-        ];
-
-        if (validRequestDestinations.indexOf(type) === -1) {
-            throw new Error('relations/HtmlPreloadLink: Not a valid type for the as-method: ' + type);
-        }
-
-        this._as = type;
-
-        if (this.node) {
-            this.node.setAttribute('as', type);
-        }
-    },
-
     attach: function (asset, position, adjacentRelation) {
         this.node = getPreloadLinkNode(this, asset);
 
         this.attachNodeBeforeOrAfter(position, adjacentRelation);
-        HtmlRelation.prototype.attach.call(this, asset, position, adjacentRelation);
+        HtmlResourceHint.prototype.attach.call(this, asset, position, adjacentRelation);
 
         // checking this.crossorigin requires the relation to already be attached
         if (this.crossorigin) {
@@ -207,17 +82,13 @@ extendWithGettersAndSetters(HtmlPreloadLink.prototype, {
     attachToHead: function (asset, position, adjacentNode) {
         this.node = getPreloadLinkNode(this, asset);
 
-        HtmlRelation.prototype.attachToHead.call(this, asset, position, adjacentNode);
+        HtmlResourceHint.prototype.attachToHead.call(this, asset, position, adjacentNode);
 
         // checking this.crossorigin requires the relation to already be attached
         if (this.crossorigin) {
             // fonts should always be treated as crossorigin: https://w3c.github.io/preload/#h-note6
             this.node.setAttribute('crossorigin', 'crossorigin');
         }
-    },
-
-    inline: function () {
-        throw new Error('Not implemented');
     }
 });
